overwrite present sample data functionality
