+++ package.json
@@ -27,7 +27,6 @@
     "generate:graphql-schema": "get-graphql-schema http://localhost:4000/graphql --json > docs/schema.json",
     "generate:ssl-private-key": "openssl req -x509 -nodes -days 1 -newkey rsa:2048 -keyout ./certs/key.pem -out ./certs/cert.pem -subj '/C=US/ST=State/L=City/O=Organization/CN=localhost'",
     "import:sample-data": "tsx ./src/utilities/loadSampleData.ts",
-    "reset-import:sample-data": "tsx ./src/utilities/resetDb.ts",
     "import:sample-data:prod": "node ./build/utilities/loadSampleData.js",
     "gen:schema": "graphql-inspector introspect ./src/typeDefs/**/**/*.ts --write ./schema.graphql ",
     "update:toc": "node scripts/githooks/update-toc.js"
